more cleanup of unread notifications
