+++ src/components/content/allergies/unread-allergies-notification.tsx
@@ -1,29 +1,11 @@
-import { useEffect, useState } from "react";
-import { useCTW } from "@/components/core/providers/use-ctw";
 import { UnreadNotification } from "@/components/core/unread-notification";
 import { usePatientAllergies } from "@/fhir/allergies";
 
-export const UnreadAllergiesNotification = () => {
+export type UnreadAllergiesNotificationProps = {
+  className?: string;
+};
+
+export const UnreadAllergiesNotification = ({ className }: UnreadAllergiesNotificationProps) => {
   const query = usePatientAllergies();
-  const { getRequestContext } = useCTW();
-
-  const [userBuilderId, setUserBuilderId] = useState("");
-
-  useEffect(() => {
-    async function load() {
-      const requestContext = await getRequestContext();
-      setUserBuilderId(requestContext.builderId);
-    }
-
-    void load();
-  }, [getRequestContext]);
-
-  const unreadOutsideAllergies = query.data.filter(
-    (allergy) => !allergy.isDismissed && !allergy.isRead && !allergy.ownedByBuilder(userBuilderId)
-  );
-
-  if (unreadOutsideAllergies.length > 0) {
-    return <UnreadNotification />;
-  }
-  return null;
+  return <UnreadNotification className={className} query={query} />;
 };
